+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - test: add `UBO_TEST_INVESTIGATION_MODE` to enable advanced and costly tools like recording stack-trace of `cell`s, generating dependency graph using `objgraph` and running pdb session when memory leak is found to better investigate memory leaks
 - test: move all dbus custom interfaces in `ubo_app.utils.dbus_interfaces` and preserve it in test cleanup, this is due to sdbus having a hidden mapping of dbus interfaces to their implementations in the C code and it doesn't clean up the mapping when the interface is removed
 - refactor: avoid using `Clock.schedule_interval` and replace it with `asyncio.sleep` in services needing to periodically run a function
+- test: add `test_menu`, as the first in a category of tests purposed to reproduce rare and hard-to-reproduce bugs: these tests run a few times normally, but when `UBO_TEST_INVESTIGATION_MODE` environment variable is set, they repeat the expected reproduction steps thousands of times until the bug is reproduced, and then run a pdb session for investigation
 
 ## Version 1.2.2